+++ tests/wallet/test_wallet_test_framework.py
@@ -3,7 +3,7 @@ from __future__ import annotations
 import pytest
 
 from chia.wallet.cat_wallet.cat_wallet import CATWallet
-from tests.wallet.conftest import BalanceCheckingError, WalletEnvironment, WalletStateTransition, WalletTestFramework
+from tests.environments import BalanceCheckingError, WalletEnvironment, WalletStateTransition, WalletTestFramework
 
 
 @pytest.mark.parametrize(
